seperate styles add constants
