+++ frontend/app/(tabs)/index.tsx
@@ -1,6 +1,8 @@
 import { StyleSheet, View, Text } from "react-native";
 import { Stack } from "expo-router";
 
+import styles from "@/constants/constants";
+
 export default function TabOneScreen() {
   return (
     <View>
@@ -9,20 +11,3 @@ export default function TabOneScreen() {
     </View>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    alignItems: "center",
-    justifyContent: "center",
-  },
-  title: {
-    fontSize: 20,
-    fontWeight: "bold",
-  },
-  separator: {
-    marginVertical: 30,
-    height: 1,
-    width: "80%",
-  },
-});
