+++ frontend/constants/constants.tsx
@@ -1,0 +1,3 @@
+import styles from "@/styles/styles";
+
+export default styles;
